// packages/lib/src/resources/application-gateway/types.ts
export enum ApplicationGatewaySkuName {
  Standard_v2 = 'Standard_v2',
  WAF_v2 = 'WAF_v2',
}

/**
 * Application Gateway SKU tier.
 */
export enum ApplicationGatewayTier {
  Standard_v2 = 'Standard_v2',
  WAF_v2 = 'WAF_v2',
}

/**
 * Application Gateway protocol.
 */
export enum ApplicationGatewayProtocol {
  Http = 'Http',
  Https = 'Https',
}